Fixed more problems related to recent restructuring of glue code generation to support simultaneous 32- and 64-bit ports. Now uses size() (on Java side) and sizeof (on C side) when creating direct Buffers for struct accessors rather than passing MachineDescription down to JavaMethodBindingEmitter and CMethodBindingEmitter. Now uses 64-bit MachineDescription when sizing outgoing arguments and return types for bound methods.
git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@506 232f8b59-042b-4e1e-8c03-345bb8c30851

+++ src/classes/com/sun/opengl/impl/windows/WindowsGLDrawable.java
@@ -199,7 +199,7 @@ public abstract class WindowsGLDrawable extends GLDrawableImpl {
               }
               gotAvailableCaps = true;
             } else {
-              int lastErr = WGL.GetLastError();
+              long lastErr = WGL.GetLastError();
               // Intel Extreme graphics fails with a zero error code
               if (lastErr != 0) {
                 throw new GLException("Unable to enumerate pixel formats of window using wglGetPixelFormatAttribivARB: error code " + WGL.GetLastError());
@@ -262,7 +262,7 @@ public abstract class WindowsGLDrawable extends GLDrawableImpl {
       pixelFormat = WGL.ChoosePixelFormat(hdc, pfd);
     }
     if (!WGL.SetPixelFormat(hdc, pixelFormat, pfd)) {
-      int lastError = WGL.GetLastError();
+      long lastError = WGL.GetLastError();
       if (DEBUG) {
         System.err.println(getThreadName() + ": SetPixelFormat failed: current context = " + WGL.wglGetCurrentContext() +
                            ", current DC = " + WGL.wglGetCurrentDC());
